Enable long marker k-mers (k up to 126) by default
Add a POD 256-bit KmerId (Uint256) so dinara supports marker k-mers with
k > 64, matching shasta2's capacity-128 Kmer. Previously KmerId was a
128-bit integer capped at k=62.

Uint256 is a flat four-word struct (not boost::multiprecision) because
KmerId is memory-mapped, hashed by raw bytes, and used as a hash-table
key, all of which require a trivially copyable layout. The on-disk marker
format is unchanged (CompressedMarker stores only position), so this does
not alter serialized data.

Gated behind the DINARA_LONG_MARKERS CMake option, default ON. Turn OFF
(-DDINARA_LONG_MARKERS=OFF) for full-genome small-k runs to save transient
memory (16 vs 32 bytes per stored KmerId).

Also fixes shift UB in ShortBaseSequence set()/reverseComplementSlow(),
generalizes extractKmer to handle k-mers straddling 3 blocks, and makes
the --Kmers.k upper bound flag-aware in main.cpp.

Verified: full GIAB assembly at k=70 completes; long-markers unit tests
pass (438487 assertions); OFF build still compiles and caps at k=62.

+++ src/KmerTable.cpp
@@ -111,7 +111,7 @@ void KmerTable::createKmerTable()
     for(uint64_t kmerId=0; kmerId<kmerCount; kmerId++) {
         const Kmer kmer(kmerId, k);
         const Kmer reverseComplementedKmer = kmer.reverseComplement(k);
-        kmerTable[kmerId].reverseComplementedKmerId = KmerId16(reverseComplementedKmer.id(k));
+        kmerTable[kmerId].reverseComplementedKmerId = KmerId16(uint64_t(reverseComplementedKmer.id(k)));
     }
     for(uint64_t kmerId=0; kmerId<kmerCount; kmerId++) {
         const uint64_t reverseComplementedKmerId = kmerTable[kmerId].reverseComplementedKmerId;
@@ -325,7 +325,7 @@ KmerTable1::KmerTable1(
         ++kmerCount;
 
         // If this k-mer is palindromic, we are done.
-        if(info.reverseComplementedKmerId == kmerId) {
+        if(info.reverseComplementedKmerId == uint64_t(kmerId)) {
             continue;
         }
 
@@ -665,7 +665,7 @@ KmerTable2::KmerTable2(
         ++kmerCount;
 
         // If this k-mer is palindromic, we are done.
-        if(info.reverseComplementedKmerId == kmerId) {
+        if(info.reverseComplementedKmerId == uint64_t(kmerId)) {
             continue;
         }
 
@@ -948,7 +948,7 @@ KmerTable4::KmerTable4(
             continue;
         }
         // Only store the lower KmerId in the pair.
-        if(kmerId > kmerIdRc) {
+        if(uint64_t(kmerId) > uint64_t(kmerIdRc)) {
             continue;
         }
         if(minimumDistance[uint64_t(kmerId)].second < distanceThreshold) {
